fix(bindings): resolve type mapping regressions and sample build errors
Squash of the afternoon fix series while adding @ohos.* API bindings to the
samples:
- typeMapper: fix type mapping regressions and unify enum dedup
- samples: nullable event handler senders (CS8622); use Display.IsFoldable()
  instead of non-existent static properties; fully qualify HarmonyOS API
  types to avoid MAUI namespace conflicts; DeviceInfo.Manufacture typo
- parser: bracket-counting fix for module.json5 permissions injection

+++ HarmonyOS.Bindings/Api/Http.cs
@@ -78,7 +78,7 @@ public static unsafe partial class Http
     /// <summary>
     /// createHttpResponseCache 方法
     /// </summary>
-    public static IntPtr CreateHttpResponseCache(IntPtr cacheSize)
+    public static IntPtr CreateHttpResponseCache(double cacheSize)
     {
         return NodeApi.CallMethod<IntPtr>(Module, _createHttpResponseCache, cacheSize);
     }
